Change routing from query string to path parameters
