+++ src/data/sidebarMock.ts
@@ -4,8 +4,7 @@ import { RxDotFilled as Dot } from 'react-icons/rx'
 const sidebarCategories = [
   {
     id: 3,
-    path: 'compare',
-    params: '/infrastructure',
+    path: '/compare',
     Icon: Folder,
     text: 'Infrastructure',
     showIcon: true,
@@ -15,7 +14,6 @@ const sidebarCategories = [
   {
     id: 4,
     path: '/compare',
-    params: 'dev-tools',
     Icon: Folder,
     text: 'Dev Tools',
     showIcon: true,
@@ -24,7 +22,6 @@ const sidebarCategories = [
   {
     id: 5,
     path: '/details',
-    params: 'vercel/next.js',
     text: 'vercel/next.js',
     Icon: Dot,
     showIcon: true,
@@ -34,7 +31,6 @@ const sidebarCategories = [
   {
     id: 6,
     path: '/details',
-    params: 'microsoft/playwright',
     Icon: Dot,
     text: 'microsoft/playwright',
     showIcon: true,
@@ -44,7 +40,6 @@ const sidebarCategories = [
   {
     id: 7,
     path: '/compare',
-    params: 'machine-learning',
     Icon: Folder,
     text: 'Machine Learning',
     showIcon: true,
@@ -53,7 +48,6 @@ const sidebarCategories = [
   {
     id: 8,
     path: '/compare',
-    params: 'artificial-intelligence',
     Icon: Folder,
     text: 'Artificial Intelligence',
     showIcon: true,
