+++ src/pages/compare/[id].tsx
@@ -1,27 +1,25 @@
 import { useReactTable, getCoreRowModel } from '@tanstack/react-table'
 import { FiChevronDown } from 'react-icons/fi'
 import { AiOutlinePlus } from 'react-icons/ai'
+import columns from '@/components/page/repositoryTable/columns'
 import TopBar from '@/components/page/repositoryTable/TopBar'
 import Table from '@/components/page/repositoryTable/Table'
 import FilterBar from '@/components/page/repositoryTable/Filterbar'
 import Chart from '@/components/page/details/Chart'
 import Page from '@/components/side-effects/Page'
 import Button from '@/components/pure/Button'
-import columns from '@/components/page/repositoryTable/columns'
-import respositoriesMock from '@/data/repositoriesMock'
 import { data } from '@/data/detailPageMocks'
-import { useRouter } from 'next/router'
+import respositoriesMock from '@/data/repositoriesMock'
 
 const nullFunc = () => null
 
+// @TODO Get id from url, fetch data from API & display real data
 const Compare = () => {
   const table = useReactTable({
     data: respositoriesMock,
     columns,
     getCoreRowModel: getCoreRowModel()
   })
-  const router = useRouter()
-  const { parameters } = router.query
 
   return (
     <Page>
@@ -33,7 +31,8 @@ const Compare = () => {
         <div className="flex flex-row items-center justify-between px-6 py-3.5">
           <div className="flex flex-col">
             <p className="text-12 font-medium uppercase text-gray-400">Compare</p>
-            <h1 className="text-24 font-medium">{parameters}</h1>
+            {/* @TODO Update page title */}
+            <h1 className="text-24 font-medium">Infrastructure</h1>
           </div>
           <div>
             <Button
